Preliminary support of OCI containers format (#110)
* package: add support for OCI format

Support creating and managing container images in OCI format with
podman. Container images are not published in a standard OCI registry,
they are simply published as OCI archives (ie. tar files) in a raw
POSIX directory designed to be exported by HTTP server.

Multi-arch container images build are supported. However, this first
implemention comes without static analysic of Containerfile and without
OCI archives digital signature.

* CI: setup new{uid,gid}map for podman

Setup all requirements to test Rift's OCI container features.

* Bump copyright year for OCI support

* Add missing docstrings in OCI methods and tests

+++ tests/TestUtils.py
@@ -254,7 +254,8 @@ class RiftProjectTestCase(RiftTestCase):
         for buildfile in self.buildfiles.values():
             os.unlink(buildfile)
         for src in self.pkgsrc.values():
-            os.unlink(src)
+            if os.path.exists(src):
+                os.unlink(src)
         for pkgdir in self.pkgdirs.values():
             info_path = os.path.join(pkgdir, 'info.yaml')
             if os.path.exists(info_path):
@@ -306,12 +307,12 @@ class RiftProjectTestCase(RiftTestCase):
         src_top_dir=None,
         tests=None,
     ):
-        # By default, make package in RPM format
+        # By default, make package in all supported formats
         if formats is None:
-            formats = ['rpm']
+            formats = ['rpm', 'oci']
         # Check provide package formats are supported
         for _format in formats:
-            assert(_format in ['rpm'])
+            assert(_format in ['rpm', 'oci'])
         # Set default source top dir name
         if src_top_dir is None:
             src_top_dir = f"{name}-{version}"
@@ -349,6 +350,10 @@ class RiftProjectTestCase(RiftTestCase):
                 nfo.write("    variants:\n")
                 for variant in variants:
                     nfo.write(f"    - {variant}\n")
+            if 'oci' in formats:
+                nfo.write("    oci:\n")
+                nfo.write(f"        version: '{version}'\n")
+                nfo.write(f"        release: '{release}'\n")
 
         # ./packages/pkg/pkg.spec
         if 'rpm' in formats:
@@ -367,6 +372,13 @@ class RiftProjectTestCase(RiftTestCase):
                     )
                 )
             self.buildfiles[f"{name}:rpm"] = buildfile
+
+        # ./packages/pkg/Containerfile
+        if 'oci' in formats:
+            buildfile = os.path.join(self.pkgdirs[name], 'Containerfile')
+            with open(buildfile, "w") as fh:
+                fh.write('FROM debian:stable')
+            self.buildfiles[f"{name}:oci"] = buildfile
 
         # ./packages/pkg/sources
         srcdir = os.path.join(self.pkgdirs[name], 'sources')
